restaructure code
